perf doc: Define man page date when using asciidoctor
Bugzilla: https://bugzilla.redhat.com/2233483

upstream
========
commit 21a165133c85d072051aa214099ad46a49239c46
Author: Ben Hutchings <benh@debian.org>
Date: Sat May 13 22:37:24 2023 +0200

description
===========
When building perf documentation with asciidoc, we use "git log" to
find the last commit date of each doc source and pass that to asciidoc
to use as the man page date.

When using asciidoctor, however, the current date is always used
instead.  Defining perf_date like we do for asciidoc also doesn't
work because we're not using DocBook as an intermediate format.
The asciidoctor man page backend looks for the variable "docdate",
so set that instead.

    Signed-off-by: Ben Hutchings <benh@debian.org>
    Acked-by: Ian Rogers<irogers@google.com>
    Cc: Ingo Molnar <mingo@redhat.com>
    Cc: Peter Zijlstra <peterz@infradead.org>
    Cc: Salvatore Bonaccorso <carnil@debian.org>
    Link: https://lore.kernel.org/r/ZF/1BOahN/i6xbBx@decadent.org.uk
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>