perf doc: Add support for KBUILD_BUILD_TIMESTAMP
Bugzilla: https://bugzilla.redhat.com/2233483

upstream
========
commit 61b3d2107d2aae5dc86bbbbd0c9f6151d694cdc4
Author: Ben Hutchings <benh@debian.org>
Date: Sat May 13 22:37:43 2023 +0200

description
===========
When building man pages from a Git checkout, we consistently set the
man page date based on when the input was last changed.  Otherwise, it
defaults to the build time, which is not reproducible.

Allow the date to be set through the KBUILD_BUILD_TIMESTAMP variable,
as for timestamps in the kernel itself.

    Signed-off-by: Ben Hutchings <benh@debian.org>
    Acked-by: Ian Rogers<irogers@google.com>
    Cc: Ingo Molnar <mingo@redhat.com>
    Cc: Peter Zijlstra <peterz@infradead.org>
    Cc: Salvatore Bonaccorso <carnil@debian.org>
    Link: https://lore.kernel.org/r/ZF/1F1P+b9qZ/vVH@decadent.org.uk
    Signed-off-by: Arnaldo Carvalho de Melo <acme@redhat.com>

Signed-off-by: Michael Petlan <mpetlan@redhat.com>