documentation: Convert to using Jira instead of Bugzilla
As of Sept 4, 2023, RHEL no longer uses Bugzilla and now uses Jira.

Update the documentation for Jira.

Note: This is the first in a short series of changesets to update the
documentation.  Additional changes are planned.  For example,
content/docs/bz_lifecycle.adoc has not been dropped, nor has a
Jira-focused replacement been written.  One will be provided in the next
few weeks.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>