Revert "sched: Consider task_struct::saved_state in wait_task_inactive()."
JIRA: https://issues.redhat.com/browse/RHEL-1536
Upstream status: RHEL only
Conflicts: A later patch renamed task_running() to task_on_cpu() so this
did not revert cleanly. In addition match_state does not need to be checked
for 0 due to f9fc8cad9728 sched: Add TASK_ANY for wait_task_inactive().

This reverts commit 3673cc2e6172fceeb54ddca2dd764e7e54fd3328.

This is commit a015745ca41f from the RT tree merge. It will be re-applied in
the form it was in when merged to Linus' tree as 1c06918788.

Signed-off-by: Phil Auld <pauld@redhat.com>